Add on-behalf-of creation to DevCenter 2025-04-01-preview To support on-behalf-of-creation, we need new operations and fields for the 2025-04-01 data plane API.
CreatedByUserObjectId:
In a create-on-behalf of scenario, we wish to store and expose who requested the Dev Box to be created.

DevBoxProvisioningState:PendingApproval
When a Dev Box is put in create-on-behalf-of scenarios, we do not wish to create the CPC VM until it has been approved.
This is done for security, to minimize the risk of attackers creating infected Dev Boxes which users can connect to unknowingly.

DevBox:Approve
Exposes a new action to approve the creation of a Dev Box, which was created using on-behalf-of
When calling approve, we will run the long running operation to create the CPC VM.

Put DevBox 202 Accepted:
When putting a Dev Box on behalf-of a different user, return 202 accepted if it needs approval before the long running operations to create the CPC VM can run.

/projects/{projectName}/approvals:
Lists all pending approvals within the project.
This allows project administrators to get an overview of which approvals are still pending.